Make mrb_int pointer-width (intptr_t) for 32-bit embedded targets
mrb_int was hardcoded int64_t. Switch it to intptr_t, decided at compile
time via INTPTR_MAX: int64_t on 64-bit hosts (PCs -- zero behavior change)
and int32_t on 32-bit embedded targets, where it gives native-word
arithmetic, half the memory for every integer / array element / hash slot,
and a pointer-width sp_RbVal union (the `i` and `p` members now match, no
8-vs-4 split). Pointers already travel through `.v.p`, never `.v.i`, so
narrowing `.v.i` to pointer width is safe.

Two width-sensitive constants follow suit:
  - SP_INT_NIL: INT64_MIN -> INTPTR_MIN. INT64_MIN cast to int32 is 0,
    which would make the int? nil sentinel collide with the integer 0 --
    a silent catastrophe on 32-bit. INTPTR_MIN is INT64_MIN on 64-bit
    (unchanged) and INT32_MIN on 32-bit (a representable Integer there,
    so the sentinel can collide with -2147483648, an accepted 32-bit
    limitation).
  - sp_core.c's Integer#round/ceil/floor 10^p overflow guards: the
    p>=19 cap and INT64_MAX/MIN clamps become SP_INT_POW10_LIMIT (19 on
    64-bit, 10 on 32-bit) and INTPTR_MAX/MIN. sp_core.h/.c helper ABI
    moves int64_t -> intptr_t to match mrb_int.

Verified both paths: 64-bit gates unchanged (679 tests, optcarrot 59662,
bootstrap fixpoint -- intptr==int64 makes it a no-op); and a -m32 build
runs correctly with mrb_int=int32 (arrays/strings/hash/Time/interpolation),
int32 overflow raises RangeError as it should (1000000*1000000), and
SP_INT_NIL is INT32_MIN (not 0).

sp_Time.tv_sec / bigint stay int64_t deliberately (they hold values wider
than a 32-bit Integer). The narrowing of those foreign 64-bit values to a
32-bit Integer at the Ruby boundary (Time#to_i, Integer(huge_str)) is a
follow-up: promote-mode keeps the value (bigint), otherwise RangeError,
never a silent truncate.

+++ lib/sp_core.h
@@ -2,7 +2,7 @@
  * libspinel_rt.a so they are compiled once rather than re-parsed and
  * re-codegen'd in every generated translation unit.
  *
- * Signatures use int64_t / double directly (== mrb_int / mrb_float)
+ * Signatures use intptr_t / double directly (== mrb_int / mrb_float)
  * to stay decoupled from the runtime's typedefs and the mrb_bool
  * conflict between sp_runtime.h and mruby_shim.h -- same convention as
  * sp_time.h. These helpers call sp_raise_cls / sp_sprintf, which live
@@ -14,24 +14,24 @@
 #include <stdint.h>
 
 /* String -> number parsers (cold, I/O-boundary). */
-int64_t sp_str_to_i_cruby(const char *s);
-int64_t sp_str_to_i_base(const char *s, int64_t base);
-int64_t sp_str_to_i_strict(const char *s);
-int64_t sp_str_to_i_strict_base(const char *s, int64_t base);
+intptr_t sp_str_to_i_cruby(const char *s);
+intptr_t sp_str_to_i_base(const char *s, intptr_t base);
+intptr_t sp_str_to_i_strict(const char *s);
+intptr_t sp_str_to_i_strict_base(const char *s, intptr_t base);
 double  sp_str_to_f_strict(const char *s);
 
 /* Cold integer-math + String#oct helpers. */
-int64_t sp_gcd(int64_t a, int64_t b);
-int64_t sp_lcm(int64_t a, int64_t b);
-int64_t sp_powmod(int64_t base, int64_t exp, int64_t mod);
-int64_t sp_ceildiv(int64_t a, int64_t b);
-int64_t sp_int_clamp(int64_t v, int64_t lo, int64_t hi);
-int64_t sp_int_sqrt(int64_t n);
-int64_t sp_ipow10(int64_t p);
-int64_t sp_int_round(int64_t v, int64_t nd);
-int64_t sp_int_ceil(int64_t v, int64_t nd);
-int64_t sp_int_floor(int64_t v, int64_t nd);
-int64_t sp_int_truncate(int64_t v, int64_t nd);
-int64_t sp_str_oct(const char *s);
+intptr_t sp_gcd(intptr_t a, intptr_t b);
+intptr_t sp_lcm(intptr_t a, intptr_t b);
+intptr_t sp_powmod(intptr_t base, intptr_t exp, intptr_t mod);
+intptr_t sp_ceildiv(intptr_t a, intptr_t b);
+intptr_t sp_int_clamp(intptr_t v, intptr_t lo, intptr_t hi);
+intptr_t sp_int_sqrt(intptr_t n);
+intptr_t sp_ipow10(intptr_t p);
+intptr_t sp_int_round(intptr_t v, intptr_t nd);
+intptr_t sp_int_ceil(intptr_t v, intptr_t nd);
+intptr_t sp_int_floor(intptr_t v, intptr_t nd);
+intptr_t sp_int_truncate(intptr_t v, intptr_t nd);
+intptr_t sp_str_oct(const char *s);
 
 #endif
